Разработка на Python/less_9/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
psycopg2-binary==2.9.9
loguru==0.7.2
requests==2.31.0